+++ Testing/Code/BasicFilters/itkBilateralImageFilterTest3.cxx
@@ -111,7 +111,7 @@ int itkBilateralImageFilterTest3(int ac, char** av)
   // compare the two images
   itk::ImageRegionIterator<myImage> it(filter3->GetOutput(),filter3->GetOutput()->GetBufferedRegion());
   itk::ImageRegionIterator<myImage> rit(baseline->GetOutput(),baseline->GetOutput()->GetBufferedRegion());
-   int status = 0;
+   unsigned long status = 0;
    while (!it.IsAtEnd())
      {
      if (it.Get() != rit.Get())
